Move hooks folder + remove shareExplore
