+++ src/factory/index.js
@@ -1,2 +1,2 @@
 export { default as ChartFactory } from './ChartFactory';
-export { default as useProfileLoader } from './hooks/useProfileLoader';
+export { default as useProfileLoader } from './useProfileLoader';
